flask-firestore/requirements.txt
Flask
gunicorn
google-cloud-firestore
